# Feat/performance optimizations