--- a/index.html
+++ b/index.html
@@ -31,6 +31,9 @@
   <script src="https://unpkg.com/@tailwindcss/browser@4"></script>
   <link rel="stylesheet" href="/css/custom.css">
   
+  <!-- Performance Optimization Script -->
+  <script src="/js/performance.js" defer></script>
+  
   <!-- Advanced Analytics Script -->
   <script src="/js/analytics.js" defer></script>
   
@@ -114,7 +117,7 @@
           <!-- Date Column -->
           <div class="flex-shrink-0 w-24 pt-1">
             <time class="text-sm font-serif" style="color: var(--warm-text-muted);">
-              Jun 11
+              2025-06-11
             </time>
           </div>
           
@@ -158,7 +161,7 @@
           <!-- Date Column -->
           <div class="flex-shrink-0 w-24 pt-1">
             <time class="text-sm font-serif" style="color: var(--warm-text-muted);">
-              Jun 10
+              2025-06-10
             </time>
           </div>
           
@@ -201,7 +204,7 @@
           <!-- Date Column -->
           <div class="flex-shrink-0 w-24 pt-1">
             <time class="text-sm font-serif" style="color: var(--warm-text-muted);">
-              May 30
+              2025-05-30
             </time>
           </div>