--- a/scripts/utils/__mocks__/logger.js
+++ b/scripts/utils/__mocks__/logger.js
@@ -1,0 +1,12 @@
+const logger = {
+  debug: jest.fn(),
+  info: jest.fn(),
+  warn: jest.fn(),
+  error: jest.fn(),
+  success: jest.fn(),
+  section: jest.fn(),
+  timer: jest.fn(() => jest.fn()), // Return a mock function that can be called
+  log: jest.fn()
+};
+
+module.exports = logger;