… docs/PERFORMANCE_IMPLEMENTATION.md
… docs/PERFORMANCE_OPTIMIZATION_PLAN.md
… docs/PERFORMANCE_SUMMARY.md
… scripts/performance-monitor.js
… src/static/js/performance.js
… templates/base.html
… tests/build-site.test.js
… tests/performance-optimizations.test.js